requirements.txt
requests==2.28.2
discord.py==1.7.3